refactor: mongock external system api
