+++ community/flamingock-auditstore-mongodb-sync/build.gradle.kts
@@ -4,7 +4,7 @@ dependencies {
     implementation(project(":utils:mongodb-util"))
     implementation(project(":core:flamingock-core"))
 
-    api(project(":core:target-systems:mongodb-sync-target-system"))
+    api(project(":core:target-systems:mongodb-external-system-api"))
 //    api(project(":community:flamingock-community"))
 
     compileOnly("org.mongodb:mongodb-driver-sync:4.0.0")
@@ -12,6 +12,8 @@ dependencies {
 
     testImplementation(project(":utils:test-util"))
     testImplementation(project(":core:target-systems:nontransactional-target-system"))
+    testImplementation(project(":core:target-systems:mongodb-sync-target-system"))
+
     testImplementation(project(":e2e:core-e2e"))
     testImplementation("org.testcontainers:testcontainers-mongodb:2.0.2")
     testImplementation("org.testcontainers:testcontainers-junit-jupiter:2.0.2")
