+++ community/flamingock-auditstore-mongodb-sync/src/main/java/io/flamingock/store/mongodb/sync/internal/MongoDBSyncAuditor.java
@@ -27,11 +27,11 @@ import io.flamingock.internal.common.core.audit.AuditEntry;
 import io.flamingock.internal.common.core.audit.AuditReader;
 import io.flamingock.internal.common.mongodb.CollectionInitializator;
 import io.flamingock.internal.common.mongodb.MongoDBAuditMapper;
+import io.flamingock.internal.common.mongodb.MongoDBSyncCollectionHelper;
+import io.flamingock.internal.common.mongodb.MongoDBSyncDocumentHelper;
 import io.flamingock.internal.core.external.store.audit.LifecycleAuditWriter;
 import io.flamingock.internal.util.Result;
 import io.flamingock.internal.util.log.FlamingockLoggerFactory;
-import io.flamingock.targetsystem.mongodb.sync.util.MongoDBSyncCollectionHelper;
-import io.flamingock.targetsystem.mongodb.sync.util.MongoDBSyncDocumentHelper;
 import org.bson.Document;
 import org.bson.conversions.Bson;
 import org.slf4j.Logger;
