+++ community/flamingock-auditstore-mongodb-sync/src/main/java/io/flamingock/store/mongodb/sync/internal/MongoDBSyncLockService.java
@@ -26,8 +26,8 @@ import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.Filters;
 import com.mongodb.client.model.UpdateOptions;
 import com.mongodb.client.result.UpdateResult;
-import io.flamingock.targetsystem.mongodb.sync.util.MongoDBSyncCollectionHelper;
-import io.flamingock.targetsystem.mongodb.sync.util.MongoDBSyncDocumentHelper;
+import io.flamingock.internal.common.mongodb.MongoDBSyncCollectionHelper;
+import io.flamingock.internal.common.mongodb.MongoDBSyncDocumentHelper;
 import io.flamingock.internal.util.id.RunnerId;
 import io.flamingock.internal.util.TimeService;
 import io.flamingock.internal.core.external.store.lock.community.CommunityLockService;
