+++ core/target-systems/mongodb-sync-target-system/build.gradle.kts
@@ -4,7 +4,7 @@ dependencies {
     api(project(":core:flamingock-core"))
     implementation(project(":utils:mongodb-util"))
     implementation(project(":legacy:mongock-importer-mongodb"))
-    api(project(":core:target-systems:mongodb-target-system-api"))
+    api(project(":core:target-systems:mongodb-external-system-api"))
 
     //General
     compileOnly("org.mongodb:mongodb-driver-sync:4.0.0")
