+++ core/target-systems/mongodb-sync-target-system/src/main/java/io/flamingock/targetsystem/mongodb/sync/MongoDBSyncTargetSystemAuditMarker.java
@@ -24,8 +24,8 @@ import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.Filters;
 import io.flamingock.internal.core.transaction.TransactionManager;
 import io.flamingock.internal.util.constants.CommunityPersistenceConstants;
-import io.flamingock.targetsystem.mongodb.sync.util.MongoDBSyncCollectionHelper;
-import io.flamingock.targetsystem.mongodb.sync.util.MongoDBSyncDocumentHelper;
+import io.flamingock.internal.common.mongodb.MongoDBSyncCollectionHelper;
+import io.flamingock.internal.common.mongodb.MongoDBSyncDocumentHelper;
 import io.flamingock.internal.common.cloud.vo.TargetSystemAuditMarkType;
 import io.flamingock.internal.common.mongodb.CollectionInitializator;
 import io.flamingock.internal.core.external.targets.mark.TargetSystemAuditMark;
