+++ core/target-systems/mongodb-sync-target-system/src/test/java/io/flamingock/targetsystem/mongodb/sync/MongoDBTestHelper.java
@@ -18,8 +18,8 @@ package io.flamingock.targetsystem.mongodb.sync;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import io.flamingock.internal.util.constants.CommunityPersistenceConstants;
-import io.flamingock.targetsystem.mongodb.sync.util.MongoDBSyncCollectionHelper;
-import io.flamingock.targetsystem.mongodb.sync.util.MongoDBSyncDocumentHelper;
+import io.flamingock.internal.common.mongodb.MongoDBSyncCollectionHelper;
+import io.flamingock.internal.common.mongodb.MongoDBSyncDocumentHelper;
 import io.flamingock.internal.common.cloud.vo.TargetSystemAuditMarkType;
 import io.flamingock.internal.core.external.targets.mark.TargetSystemAuditMark;
 import io.flamingock.internal.core.external.store.audit.domain.AuditContextBundle;
