+++ utils/mongodb-util/src/main/java/io/flamingock/internal/common/mongodb/MongoDBSyncDocumentHelper.java
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.flamingock.targetsystem.mongodb.sync.util;
+package io.flamingock.internal.common.mongodb;
 
-import io.flamingock.internal.common.mongodb.DocumentHelper;
 import org.bson.Document;
 
 public class MongoDBSyncDocumentHelper implements DocumentHelper {
